refactor: move voucher and json response code into framework
