+++ gateway-framework/Cargo.toml
@@ -7,6 +7,7 @@ version = "0.0.1"
 alloy-primitives.workspace = true
 alloy-sol-types = "0.5.3"
 anyhow.workspace = true
+axum.workspace = true
 ethers = "2.0.11"
 eventuals = "0.6.7"
 gateway-common = { path = "../gateway-common" }
