+++ gateway-framework/src/json.rs
@@ -1,0 +1,20 @@
+use std::iter;
+
+use axum::Json;
+use reqwest::header::{self, HeaderMap, HeaderName, HeaderValue};
+
+pub type JsonResponse = (HeaderMap, Json<serde_json::Value>);
+
+pub fn json_response<H>(headers: H, payload: serde_json::Value) -> JsonResponse
+where
+    H: IntoIterator<Item = (HeaderName, HeaderValue)>,
+{
+    let headers = HeaderMap::from_iter(
+        iter::once((
+            header::CONTENT_TYPE,
+            HeaderValue::from_static("application/json"),
+        ))
+        .chain(headers),
+    );
+    (headers, Json(payload))
+}
