+++ gateway-framework/src/scalar/mod.rs
@@ -1,3 +1,5 @@
 mod receipts;
+mod vouchers;
 
 pub use receipts::{ReceiptSigner, ReceiptStatus, ScalarReceipt};
+pub use vouchers::{handle_collect_receipts, handle_partial_voucher, handle_voucher};
