+++ graph-gateway/src/lib.rs
@@ -1,8 +1,3 @@
-use std::iter;
-
-use axum::Json;
-use reqwest::header::{self, HeaderMap, HeaderName, HeaderValue};
-
 pub mod auth;
 pub mod block_constraints;
 pub mod budgets;
@@ -17,20 +12,3 @@ pub mod subscriptions;
 pub mod subscriptions_subgraph;
 pub mod topology;
 pub mod unattestable_errors;
-pub mod vouchers;
-
-pub type JsonResponse = (HeaderMap, Json<serde_json::Value>);
-
-pub fn json_response<H>(headers: H, payload: serde_json::Value) -> JsonResponse
-where
-    H: IntoIterator<Item = (HeaderName, HeaderValue)>,
-{
-    let headers = HeaderMap::from_iter(
-        iter::once((
-            header::CONTENT_TYPE,
-            HeaderValue::from_static("application/json"),
-        ))
-        .chain(headers),
-    );
-    (headers, Json(payload))
-}
